Add reverse ordering and update in place

+++ app/helpers/message_three.rb
@@ -9,7 +9,12 @@ def build_timer_message_three(timers: nil)
 
 
   timers ||= Timer.all
-  timers.sort_by {|timer| next_spawn_time_start(timer.name, timer: timer) || Chronic.parse("100 years from now") }.each do |timer|
+  timers = timers.sort_by {|timer| next_spawn_time_start(timer.name, timer: timer) || Chronic.parse("100 years from now") }
+  if TIMER_MESSAGE_THREE_ORDER == "reverse"
+    timers = timers.reverse
+  end
+
+  timers.each do |timer|
     window_start = ""
     vague_window_start = ""
     starts_at = ""
@@ -105,54 +110,70 @@ def build_timer_message_three(timers: nil)
   client = Discordrb::Webhooks::Client.new(url: TIMER_CHANNEL_WEBHOOK_URL)
   builder = Discordrb::Webhooks::Builder.new
   builder.content = ""
-  builder.add_embed do |embed|
+
+  embeds = []
+  embeds << Proc.new {|embed|
     embed.color = any_in_window ? 15105570 : 3066993
     embed.title = any_in_window ? "Mobs In Window" : "Nothing Currently in Window"
     embed.fields = mobs_in_window.map {|m| m[:message] }
     embed.footer =  Discordrb::Webhooks::EmbedFooter.new(text: any_in_window ? "These are currently in window! Be prepared! • Today at #{Time.now.strftime("%I:%M:%S %p")}" : "There is currently nothing in window! • Today at #{Time.now.strftime("%I:%M:%S %p")}")
-  end
-  builder.add_embed do |embed|
+  }
+
+  embeds << Proc.new {|embed|
     embed.color = 3447003
     embed.title = "Mobs Entering Window In The Next 24 Hours"
     embed.fields = upcoming_window
-  end
+  }
+
   if SHOW_FUTURE_WINDOW && future_window.size > 0
-    builder.add_embed do |embed|
+    embeds << Proc.new {|embed|
       embed.title = "Future Windows"
       embed.description = future_window.join("\n")
     end
   end
 
-  webhook_message_id = Setting.find_by_key("webhook_message_id")
-  if webhook_message_id
-    begin
-      channel = BOT.channel(TIMER_CHANNEL_ID)
-      channel.delete_message(webhook_message_id)
-    rescue => ex
-    end
+  if TIMER_MESSAGE_THREE_ORDER == "reverse"
+    embeds = embeds.reverse
   end
 
-  result = client.execute(builder, true)
-  response = JSON.parse(result.body)
-  webhook_message_id = response["id"]
-  Setting.save_by_key("webhook_message_id", webhook_message_id)
+  embeds.each do |embed|
+    builder.add_embed(&embed)
+  end
 
+  # Delete and create
+  if false
+    webhook_message_id = Setting.find_by_key("webhook_message_id")
+    if webhook_message_id
+      begin
+        channel = BOT.channel(TIMER_CHANNEL_ID)
+        channel.delete_message(webhook_message_id)
+      rescue => ex
+      end
+    end
 
-  # if webhook_message_id == nil
-  #   result = client.execute(builder, true)
-  #   response = JSON.parse(result.body)
-  #   webhook_message_id = response["id"]
-  #   Setting.save_by_key("webhook_message_id", webhook_message_id)
-  # else
-  #   begin
-  #     client.edit_message(webhook_message_id, builder: builder)
-  #   rescue => ex
-  #     if ex.message =~ /404 Not Found/
-  #       result = client.execute(builder, true)
-  #       response = JSON.parse(result.body)
-  #       webhook_message_id = response["id"]
-  #       Setting.save_by_key("webhook_message_id", webhook_message_id)
-  #     end
-  #   end
-  # end
+    result = client.execute(builder, true)
+    response = JSON.parse(result.body)
+    webhook_message_id = response["id"]
+    Setting.save_by_key("webhook_message_id", webhook_message_id)
+
+  # Create or update
+  else
+    if webhook_message_id == nil
+      result = client.execute(builder, true)
+      response = JSON.parse(result.body)
+      webhook_message_id = response["id"]
+      Setting.save_by_key("webhook_message_id", webhook_message_id)
+    else
+      begin
+        client.edit_message(webhook_message_id, builder: builder)
+      rescue => ex
+        if ex.message =~ /404 Not Found/
+          result = client.execute(builder, true)
+          response = JSON.parse(result.body)
+          webhook_message_id = response["id"]
+          Setting.save_by_key("webhook_message_id", webhook_message_id)
+        end
+      end
+    end
+  end
 end
